add main and not found pages, move todo list from app, add new paths to the router
