+++ src/index.css
@@ -25,7 +25,8 @@ a:hover {
 body {
   margin: 0;
   display: flex;
-  place-items: start;
+  justify-content: center;
+  place-items: center;
   min-width: 320px;
   min-height: 100vh;
 }
